refactored login hooks
