+++ src/App.jsx
@@ -1,8 +1,7 @@
 import { useState, useMemo, useEffect } from 'react'
-import { useGoogleLogin } from '@react-oauth/google';
-import axios from 'axios';
 import './App.css'
-import { useAuthStore } from './stores/authStore'
+import { useAuthStore } from './auth/authStore'
+import { useLogin } from './auth/useLogin'
 
 // Test for Slack Integration
 
@@ -35,7 +34,7 @@ function App() {
     ]
   ], [])
 
-  const [currentUser, loginUser, logoutUser] = useAuthStore(state => [state.currentUser, state.loginUser, state.logoutUser])
+  const [currentUser, logoutUser] = useAuthStore(state => [state.currentUser, state.loginUser, state.logoutUser])
 
   const [cards, setCards] = useState(Flashcards[0])
   const [currentCard, setCurrentCard] = useState(0)
@@ -61,21 +60,7 @@ function App() {
     setPlaying(!playing)
   }
 
-  const googleLogin = useGoogleLogin({
-    onSuccess: async tokenResponse => {
-      console.log(tokenResponse);
-      // fetching userinfo can be done on the client or the server
-      const userInfo = await axios
-        .get('https://www.googleapis.com/oauth2/v3/userinfo', {
-          headers: { Authorization: `Bearer ${tokenResponse.access_token}` },
-        })
-        .then(res => res.data);
-
-      loginUser(userInfo)
-      console.log(userInfo);
-    },
-    // flow: 'implicit', // implicit is the default
-  });
+  const googleLogin = useLogin()
 
   function handleCardClick() {
     facing === 'front' ? setFacing('back') : setFacing('front')
